+++ packages/connectors/src/exports/index.test-d.ts
@@ -1,4 +1,0 @@
-import { expectTypeOf } from 'vitest'
-
-// noop test because vitest typecheck fails unless each workspace project has type test
-expectTypeOf(1).toEqualTypeOf<number>()