+++ packages/connectors/src/metaMask.ts
@@ -1,8 +1,8 @@
 import {
   EventType,
-  MetaMaskSDK,
+  type MetaMaskSDK,
   type MetaMaskSDKOptions,
-  SDKProvider,
+  type SDKProvider,
 } from '@metamask/sdk'
 import {
   ChainNotConfiguredError,
@@ -174,9 +174,10 @@ export function metaMask(parameters: MetaMaskParameters = {}) {
     async getProvider() {
       if (!walletProvider) {
         if (!sdk || !sdk?.isInitialized()) {
+          const { MetaMaskSDK } = await import('@metamask/sdk')
           sdk = new MetaMaskSDK({
-            enableDebug: false,
             dappMetadata: { name: 'wagmi' },
+            enableAnalytics: false,
             extensionOnly: true,
             modals: {
               // Disable by default since it pops up when mobile tries to reconnect