+++ packages/react/src/exports/index.test.ts
@@ -20,6 +20,7 @@ test('exports', () => {
       "useBytecode",
       "useCall",
       "useChainId",
+      "useChains",
       "useClient",
       "useConfig",
       "useConnect",