+++ packages/register-tests/react/src/useSwitchChain.test-d.ts
@@ -1,11 +1,15 @@
 import { expectTypeOf, test } from 'vitest'
 import { useSwitchChain } from 'wagmi'
-import { mainnet } from 'wagmi/chains'
+import { type Chain, celo, mainnet, optimism } from 'wagmi/chains'
 
 import { type ChainId } from './config.js'
 
-test('default', async () => {
-  const { switchChain } = useSwitchChain()
+test('default', () => {
+  const { chains, switchChain } = useSwitchChain()
+
+  expectTypeOf(chains[0]).toEqualTypeOf<Chain | typeof celo>()
+  expectTypeOf(chains[2]).toEqualTypeOf<Chain | typeof optimism>()
+  expectTypeOf(chains[5]).toEqualTypeOf<Chain | undefined>()
 
   switchChain(
     { chainId: 1 },