+++ packages/react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # wagmi
 
+## 2.5.7
+
+### Patch Changes
+
+- [#3580](https://github.com/wevm/wagmi/pull/3580) [`c677dcd2`](https://github.com/wevm/wagmi/commit/c677dcd245dccdf69289a3d66dded237b09570a2) Thanks [@tmm](https://github.com/tmm)! - Made `useSwitchChain().chains` reactive.
+
+- Updated dependencies [[`ca5decdb`](https://github.com/wevm/wagmi/commit/ca5decdb712f81e3f5dab933a94b967bca5b6af4), [`c677dcd2`](https://github.com/wevm/wagmi/commit/c677dcd245dccdf69289a3d66dded237b09570a2)]:
+  - @wagmi/connectors@4.1.14
+  - @wagmi/core@2.6.5
+
+## 2.5.6
+
+### Patch Changes
+
+- Updated dependencies [[`7c6618e6`](https://github.com/wevm/wagmi/commit/7c6618e6a0eb1ff39cf8f66b34d3ddc14be538fe), [`fa25b448`](https://github.com/wevm/wagmi/commit/fa25b4482504b4d9729a5687ea6d6dc959265bc0), [`895f28e8`](https://github.com/wevm/wagmi/commit/895f28e873af7c8eda5ca85734ff67c8979fd950)]:
+  - @wagmi/core@2.6.4
+  - @wagmi/connectors@4.1.13
+
 ## 2.5.5
 
 ### Patch Changes